+++ Sources/Log/Classes/Helpers/Writers/ConsoleWriter/ConsoleWriter.swift
@@ -1,6 +1,6 @@
 //
 // log
-// Copyright © 2023 Space Code. All rights reserved.
+// Copyright © 2024 Space Code. All rights reserved.
 //
 
 import Foundation
@@ -8,15 +8,15 @@ import Foundation
 // MARK: - ConsoleWriter
 
 /// A class that conforms to the IConsoleWriter protocol and writes messages to the console output.
-public final class ConsoleWriter: IConsoleWriter {
+final class ConsoleWriter: IConsoleWriter {
     // MARK: Initialization
 
     /// Initializes a new ConsoleWriter instance.
-    public init() {}
+    init() {}
 
     // MARK: IConsoleWriter
 
-    public func print(_ message: String) {
+    func print(_ message: String) {
         Swift.print(message)
     }
 }